feat: add command to set current date and time

+++ include/uart_drv.h
@@ -12,7 +12,7 @@ void uart_enable_rx_interrupt(void);
 
 void uart_send_char(char c);
 void uart_send_string(const char *str);
-void uart_send_uint(uint16_t value);
+void uart_send_uint(uint32_t value);
 
 char uart_receive_char(void);
